// apps/api-gateway/src/config/broker/alice-blue/alice-blue-broker.enum.ts
export enum WebsocketState {
  DISCONNECTED = 'DISCONNECTED',
  CONNECTING = 'CONNECTING',
  AUTHENTICATING = 'AUTHENTICATING',
  CONNECTED = 'CONNECTED',
  RECONNECTING = 'RECONNECTING',
  ERROR = 'ERROR',
}

export enum WebsocketMessageType {
  CONNECTION_REQUEST = 'c',
  CONNECTION_ACK = 'cf',
  MARKET_SUBSCRIPTION_ACK = 'tk',
  MARKET_TICK = 'tf',
  DEPTH_SUBSCRIPTION_ACK = 'dk',
  DEPTH_UPDATE = 'df',
  HEARTBEAT = 'h',
}